zmiany PD na 11.03.2023 (#3)
* Dodanie formularza tworzenia Studenta

* Student form
Fixed Teacher form

* Lesson form
Need to add back custom.css and style.css

* Lesson form
Need to add back custom.css and style.css

* updated

* added exepctions and made less code.

* Trying to add teacher list to lesson html.

* Trying to add teacher list to lesson html.

* Cannot delete already started lesson method created.

* Cannot create lesson with past date created.

* added list of teachers and students in lesson

* added list of teachers and students in lesson

* added code: student cannot be created when teacher does not teach the langugage which student want to learn

* added code: student cannot be created when teacher does not teach the langugage which student want to learn

* Adding teacher to student.

* CR fixes

---------

+++ src/main/java/pl/masters/coding/lesson/LessonController.java
@@ -1,42 +1,57 @@
 package pl.masters.coding.lesson;
 
 import lombok.RequiredArgsConstructor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import pl.masters.coding.lesson.model.Lesson;
+import pl.masters.coding.student.StudentController;
+import pl.masters.coding.student.StudentRepository;
+import pl.masters.coding.student.StudentService;
+import pl.masters.coding.student.model.Student;
+import pl.masters.coding.teacher.TeacherService;
 
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.List;
+import java.util.regex.Pattern;
 
 @Controller
 @RequiredArgsConstructor
 @RequestMapping("/lessons")
 public class LessonController {
-    private static final Logger logger = LoggerFactory.getLogger(LessonController.class);
-
-    private final LessonServiceImpl lessonService;
+    private final LessonService lessonService;
+    private final TeacherService teacherService;
+    private final StudentService studentService;
 
     @GetMapping
     public String getLessonList(Model model) {
-        List<Lesson> allLessons = lessonService.findAll();
-        model.addAttribute("lessons", allLessons);
+        model.addAttribute("lessons", lessonService.findAll());
+        model.addAttribute("teachers", teacherService.findAll());
+        model.addAttribute("students", studentService.findAll());
         return "lesson/list";
     }
 
     @PostMapping("/create")
-    //elementy z formularza - przekazane do metody (widok) - modelAttribute
-    public String createLesson(@ModelAttribute Lesson lesson) {
-        lessonService.createLesson(lesson);
-        System.out.println(lesson);
+    public String createLesson(@ModelAttribute Lesson lesson, @RequestParam("datetime") String date) {
+        //TODO: do ogarnięcia przeniesienie walidacji do serwisu lub zrobienie walidacji adnotacjami na poziomie klasy (spring-boot-starter-validation)
+        if (lesson.getDate().isBefore(LocalDateTime.now())){
+            throw new IllegalArgumentException("Cannot create lesson in the past!");
+        } else {
+            lessonService.createLesson(lesson);
+        }
         return "redirect:/lessons";
     }
 
     @GetMapping("/delete")
-    public String deleteLesson(@RequestParam Long id) {
-        lessonService.deleteLesson(id);
+    public String deleteLesson(@RequestParam int id) {
+        Lesson lesson = lessonService.findLessonById(id);
+        // TODO: 3/11/2023 przeniesienie logiki walidującej na poziom serwisu
+        if (lesson.getDate().isBefore(LocalDateTime.now())) {
+            throw new IllegalArgumentException("Cannot delete lesson which already started!");
+        } else {
+            lessonService.deleteLesson(id);
+        }
         return "redirect:/lessons";
     }
-
 }
